feat: Adiciona IA para prever informações de estoque. Cosumido a API da OPENAI para utilização da IA no app

+++ MicroServicoVendas/Repositories/PedidoRepository.cs
@@ -3,6 +3,7 @@ using MicroServicoVendas.Models;
 using MicroServicoVendas.Repositories.Context;
 using Microsoft.EntityFrameworkCore;
 using System.Linq;
+using System.Threading.Tasks;
 
 namespace MicroServicoVendas.Repositories
 {
@@ -15,7 +16,7 @@ namespace MicroServicoVendas.Repositories
             _context = context;
         }
 
-        public async Task<Pedido> CreatePedido(Pedido pedido)
+        public async Task<Pedido> CreatePedidoAsync(Pedido pedido)
         {
             var novoPedido = await _context.Pedidos.AddAsync(pedido);
 
@@ -29,16 +30,16 @@ namespace MicroServicoVendas.Repositories
             return novoPedido.Entity;
         }
 
-        public List<Pedido> GetPedidos()
+        public async Task<List<Pedido>> GetPedidosAsync()
         {
-            var listaPedidos = _context.Pedidos.AsNoTracking().ToList();
+            var listaPedidos = await _context.Pedidos.AsNoTracking().ToListAsync();
 
             return listaPedidos;
         }
 
-        public Pedido GetPedido(int id)
+        public async Task<Pedido> GetPedidoAsync(int id)
         {
-            var pedido = _context.Pedidos.AsNoTracking().FirstOrDefault(p => p.Id == id);
+            var pedido = await _context.Pedidos.AsNoTracking().FirstOrDefaultAsync(p => p.Id == id);
 
             return pedido;
         }
@@ -57,6 +58,25 @@ namespace MicroServicoVendas.Repositories
             return pedido;
         }
 
+        public async Task<List<int>> GetListCountVendasMensaisByProdutoAsync(int produtoId)
+        {
+            
+            var pedidos = await
+            (
+                from p in _context.Pedidos
+                where p.ProdutoId == produtoId
+                select p
+            )
+            .AsNoTracking()
+            .ToListAsync();
+
+            var listCountMensalProduto = pedidos
+                .GroupBy(p => new { Year = p.DataPedido.Year, Month = p.DataPedido.Month })
+                .OrderBy(g => g.Key.Year).ThenBy(g => g.Key.Month)
+                .Select(g => g.Sum(p => p.Quantidade))
+                .ToList();
+            return listCountMensalProduto;
+        }
         public async Task<Pedido> UpdatePedido(int id, Pedido pedido)
         {
             var pedidoUpdate = await _context.Pedidos.AsNoTracking().FirstOrDefaultAsync(p => p.Id == id);
@@ -76,18 +96,23 @@ namespace MicroServicoVendas.Repositories
             return pedidoUpdate;
         }
 
-        public void DeletePedido(int id)
+        public async Task DeletePedidoAsync(int id)
         {
-            var pedidoDeleted = GetPedido(id);
+            var pedidoDeleted = await GetPedidoAsync(id);
+
+            if (pedidoDeleted == null)
+            {
+                throw new Exception("Pedido not found.");
+            }
+
             _context.Pedidos.Remove(pedidoDeleted);
 
-            var validateIfSaveChanges = _context.SaveChanges() > 0;
+            var validateIfSaveChanges = await _context.SaveChangesAsync() > 0;
 
             if (!validateIfSaveChanges)
             {
                 throw new Exception("Failed to save Pedido.");
             }
-            
         }
     }
 }
